fix: move file logger from Core directory to Logging directory
